refactor: Clean up codebase and reorganize documentation
- Remove unused debug components (ApiDebugPanel, AuthDebugPanel)
- Clean up CLAUDE.md project documentation
- Add API status page for system monitoring
- Reorganize documentation structure (docs/plans/ folder)
- Clean up obsolete implementation plans
- Update supplier name mapping documentation
- Remove null-based masking implementation docs

This commit consolidates the codebase by removing unused components,
reorganizing documentation, and cleaning up obsolete plans while
adding new monitoring capabilities.
