+++ app/api-status/page.tsx
@@ -3,6 +3,9 @@
 import { useState, useEffect } from 'react';
 import { Button } from '../../components/ui';
 import { Loader, Check, X, AlertTriangle } from 'lucide-react';
+import { useAuth } from '../../hooks/useAuth';
+import { apiClient } from '@/utils/api';
+import type { OAuthScope } from '@/types/product';
 
 // API status checking interface
 interface ApiStatus {
@@ -12,7 +15,23 @@ interface ApiStatus {
   error: string | null;
 }
 
+// Debug panel interfaces
+interface DebugStatus {
+  status: 'idle' | 'loading' | 'success' | 'error';
+  data?: unknown;
+  error?: string;
+  timestamp?: string;
+}
+
+interface DebugInfo {
+  baseUrl: string;
+  healthCheck: DebugStatus;
+  searchTest: DebugStatus;
+}
+
 export default function ApiStatusPage() {
+  const { authState, getAccessToken } = useAuth();
+  
   const [apiStatus, setApiStatus] = useState<ApiStatus>({
     isConnected: false,
     responseTime: null,
@@ -20,6 +39,14 @@ export default function ApiStatusPage() {
     error: null,
   });
   const [isChecking, setIsChecking] = useState(false);
+  
+  // Debug panel state
+  const [debugInfo, setDebugInfo] = useState<DebugInfo>({
+    baseUrl: process.env.NEXT_PUBLIC_EXTERNAL_API_URL || process.env.NEXT_PUBLIC_API_BASE_URL || '/api',
+    healthCheck: { status: 'loading' },
+    searchTest: { status: 'idle' }
+  });
+  const [testQuery, setTestQuery] = useState('terrassefornyer');
 
   // API endpoint check function
   const checkApiConnection = async () => {
@@ -65,9 +92,80 @@ export default function ApiStatusPage() {
     }
   };
 
+  // Debug API functions
+  const testHealthCheck = async () => {
+    setDebugInfo(prev => ({
+      ...prev,
+      healthCheck: { status: 'loading' }
+    }));
+
+    try {
+      const result = await apiClient.healthCheck();
+      setDebugInfo(prev => ({
+        ...prev,
+        healthCheck: {
+          status: 'success',
+          data: result,
+          timestamp: new Date().toISOString()
+        }
+      }));
+    } catch (error) {
+      setDebugInfo(prev => ({
+        ...prev,
+        healthCheck: {
+          status: 'error',
+          error: error instanceof Error ? error.message : 'Unknown error',
+          timestamp: new Date().toISOString()
+        }
+      }));
+    }
+  };
+
+  const testSearch = async () => {
+    setDebugInfo(prev => ({
+      ...prev,
+      searchTest: { status: 'loading' }
+    }));
+
+    try {
+      const accessToken = await getAccessToken();
+      
+      const result = await apiClient.searchProducts({
+        søketekst: testQuery.trim() || 'test',
+        side: 1,
+        sideStørrelse: 5,
+        sortering: 'relevans'
+      }, accessToken || undefined);
+      
+      setDebugInfo(prev => ({
+        ...prev,
+        searchTest: {
+          status: 'success',
+          data: {
+            ...result,
+            authStatus: accessToken ? 'authenticated' : 'anonymous',
+            tokenPresent: !!accessToken,
+            userScopes: authState.scopes || []
+          },
+          timestamp: new Date().toISOString()
+        }
+      }));
+    } catch (error) {
+      setDebugInfo(prev => ({
+        ...prev,
+        searchTest: {
+          status: 'error',
+          error: error instanceof Error ? error.message : 'Unknown error',
+          timestamp: new Date().toISOString()
+        }
+      }));
+    }
+  };
+
   // Check API on component mount
   useEffect(() => {
     checkApiConnection();
+    testHealthCheck();
   }, []);
 
   const getStatusColor = () => {
@@ -254,6 +352,219 @@ export default function ApiStatusPage() {
             </div>
           </div>
         </div>
+
+        {/* Debug Panels - Development Mode Only */}
+        {process.env.NEXT_PUBLIC_ENABLE_DEVTOOLS === 'true' && (
+          <>
+            {/* API Debug Panel */}
+            <div className="bg-white rounded-lg shadow-sm border border-neutral-200 p-6 mb-6">
+              <h2 className="text-xl font-semibold text-neutral-800 mb-4">
+                API Debug Panel
+              </h2>
+              
+              {/* Base URL */}
+              <div className="mb-4">
+                <strong>Base URL:</strong>
+                <div className="font-mono text-sm bg-neutral-50 p-2 rounded mt-1">
+                  {debugInfo.baseUrl}
+                </div>
+              </div>
+
+              {/* Health Check */}
+              <div className="mb-4">
+                <div className="flex items-center gap-2 mb-2">
+                  <strong>Health Check:</strong>
+                  <span className={`inline-block w-3 h-3 rounded-full ${
+                    debugInfo.healthCheck.status === 'loading' ? 'bg-yellow-500' :
+                    debugInfo.healthCheck.status === 'success' ? 'bg-green-500' : 'bg-red-500'
+                  }`}></span>
+                  <button 
+                    onClick={testHealthCheck}
+                    disabled={debugInfo.healthCheck.status === 'loading'}
+                    className="text-sm bg-neutral-100 hover:bg-neutral-200 px-2 py-1 rounded disabled:opacity-50"
+                  >
+                    Test
+                  </button>
+                </div>
+                
+                {debugInfo.healthCheck.error && (
+                  <div className="bg-red-50 border border-red-200 rounded p-2 text-sm">
+                    <strong>Error:</strong> {debugInfo.healthCheck.error}
+                  </div>
+                )}
+                
+                {debugInfo.healthCheck.data !== undefined && (
+                  <div className="bg-green-50 border border-green-200 rounded p-2 text-sm">
+                    <strong>Success:</strong> 
+                    <pre className="text-xs mt-1 whitespace-pre-wrap">
+                      {typeof debugInfo.healthCheck.data === 'string' 
+                        ? debugInfo.healthCheck.data 
+                        : JSON.stringify(debugInfo.healthCheck.data, null, 2)}
+                    </pre>
+                  </div>
+                )}
+              </div>
+
+              {/* Search Test */}
+              <div className="mb-4">
+                <div className="flex items-center gap-2 mb-2">
+                  <strong>Search Test:</strong>
+                  <span className={`inline-block w-3 h-3 rounded-full ${
+                    debugInfo.searchTest.status === 'loading' ? 'bg-yellow-500' :
+                    debugInfo.searchTest.status === 'success' ? 'bg-green-500' : 
+                    debugInfo.searchTest.status === 'error' ? 'bg-red-500' : 'bg-neutral-300'
+                  }`}></span>
+                </div>
+                
+                <div className="flex gap-2 mb-2">
+                  <input
+                    type="text"
+                    value={testQuery}
+                    onChange={(e) => setTestQuery(e.target.value)}
+                    placeholder="Enter search query..."
+                    className="flex-1 px-2 py-1 text-sm border border-neutral-300 rounded focus:outline-none focus:ring-1 focus:ring-byggern-primary"
+                  />
+                  <button 
+                    onClick={testSearch}
+                    disabled={debugInfo.searchTest.status === 'loading'}
+                    className="text-sm bg-byggern-primary hover:bg-byggern-primary/90 text-white px-3 py-1 rounded disabled:opacity-50"
+                  >
+                    {debugInfo.searchTest.status === 'loading' ? 'Testing...' : 'Test Search'}
+                  </button>
+                </div>
+                
+                <div className="text-xs text-neutral-600 mb-2">
+                  Query: &ldquo;{testQuery || 'test'}&rdquo; • Page: 1 • Size: 5 • Sort: relevance
+                </div>
+                
+                {debugInfo.searchTest.error && (
+                  <div className="bg-red-50 border border-red-200 rounded p-2 text-sm">
+                    <strong>Error:</strong> {debugInfo.searchTest.error}
+                  </div>
+                )}
+                
+                {debugInfo.searchTest.data !== undefined && (
+                  <div className="bg-green-50 border border-green-200 rounded p-2 text-sm max-h-32 overflow-auto">
+                    <strong>Results:</strong> 
+                    <pre className="text-xs mt-1 whitespace-pre-wrap">
+                      {typeof debugInfo.searchTest.data === 'string' 
+                        ? debugInfo.searchTest.data 
+                        : JSON.stringify(debugInfo.searchTest.data, null, 2)}
+                    </pre>
+                  </div>
+                )}
+              </div>
+
+              <div className="text-xs text-neutral-500">
+                <div>Build Mode: {process.env.NODE_ENV} | Debug Mode: {process.env.NEXT_PUBLIC_ENABLE_DEVTOOLS === 'true' ? 'Development' : 'Production'}</div>
+                <div className="mt-1">API Endpoint: {debugInfo.baseUrl}</div>
+                <div className="mt-1">Auth Status: {authState.isAuthenticated ? `Logged in (${authState.scopes.length} scopes)` : 'Anonymous'}</div>
+                {authState.isAuthenticated && (
+                  <div className="mt-1">User Scopes: {authState.scopes.join(', ') || 'None'}</div>
+                )}
+              </div>
+            </div>
+
+            {/* Auth Debug Panel */}
+            <div className="bg-white rounded-lg shadow-sm border border-neutral-200 p-6 mb-6">
+              <h2 className="text-xl font-semibold text-neutral-800 mb-4">
+                Auth Debug Panel
+              </h2>
+              
+              <div className="space-y-3">
+                {/* Authentication Status */}
+                <div className="p-2 bg-neutral-50 rounded border">
+                  <div className="text-xs font-medium text-neutral-600 mb-1">Status</div>
+                  <div className={`text-sm font-semibold ${authState.isAuthenticated ? 'text-green-600' : 'text-red-600'}`}>
+                    {authState.isLoading ? 'Loading...' : authState.isAuthenticated ? 'Authenticated' : 'Not Authenticated'}
+                  </div>
+                </div>
+
+                {/* OAuth Scopes */}
+                <div className="p-2 bg-neutral-50 rounded border">
+                  <div className="text-xs font-medium text-neutral-600 mb-1">
+                    OAuth Scopes ({authState.scopes.length})
+                  </div>
+                  {authState.scopes.length > 0 ? (
+                    <div className="space-y-1">
+                      {authState.scopes.map((scope) => (
+                        <div key={scope} className="text-xs">
+                          <span className="font-mono text-green-700 bg-green-50 px-1 rounded">
+                            {scope}
+                          </span>
+                          <div className="text-neutral-600 mt-0.5">
+                            {scope === 'varekatalog/prices' ? 'Price information (Grunnpris, Nettopris)' : 
+                             scope === 'varekatalog/inventory' ? 'Stock quantities (Lagerantall)' : 
+                             'Unknown scope'}
+                          </div>
+                        </div>
+                      ))}
+                    </div>
+                  ) : (
+                    <div className="text-xs text-neutral-500 italic">No scopes available</div>
+                  )}
+                </div>
+
+                {/* Permissions */}
+                <div className="p-2 bg-neutral-50 rounded border">
+                  <div className="text-xs font-medium text-neutral-600 mb-1">Permissions</div>
+                  <div className="space-y-1 text-xs">
+                    <div className="text-green-600">
+                      Search: ✓ (Always Public)
+                    </div>
+                    <div className={`${authState.permissions.canViewPrices ? 'text-green-600' : 'text-red-600'}`}>
+                      View Prices: {authState.permissions.canViewPrices ? '✓ (Premium)' : '✗ (Login Required)'}
+                    </div>
+                    <div className={`${authState.permissions.canViewInventory ? 'text-green-600' : 'text-red-600'}`}>
+                      View Inventory: {authState.permissions.canViewInventory ? '✓ (Premium)' : '✗ (Login Required)'}
+                    </div>
+                  </div>
+                </div>
+
+                {/* Token Info */}
+                {authState.accessToken && (
+                  <div className="p-2 bg-neutral-50 rounded border">
+                    <div className="text-xs font-medium text-neutral-600 mb-1">Access Token</div>
+                    <div className="text-xs font-mono text-neutral-700 break-all">
+                      {authState.accessToken.substring(0, 50)}...
+                    </div>
+                  </div>
+                )}
+
+                {/* Error Info */}
+                {authState.error && (
+                  <div className="p-2 bg-red-50 rounded border border-red-200">
+                    <div className="text-xs font-medium text-red-600 mb-1">Error</div>
+                    <div className="text-xs text-red-700">{authState.error}</div>
+                  </div>
+                )}
+
+                {/* Data Masking Preview */}
+                <div className="p-2 bg-blue-50 rounded border">
+                  <div className="text-xs font-medium text-neutral-600 mb-1">Data Masking Preview</div>
+                  <div className="space-y-1 text-xs">
+                    <div>
+                      Stock Quantities: {authState.permissions.canViewInventory ? 
+                        <span className="text-green-600">Visible</span> : 
+                        <span className="text-red-600">****</span>
+                      }
+                    </div>
+                    <div>
+                      Prices: {authState.permissions.canViewPrices ? 
+                        <span className="text-green-600">Visible</span> : 
+                        <span className="text-red-600">****</span>
+                      }
+                    </div>
+                  </div>
+                </div>
+
+                <div className="mt-3 pt-2 border-t text-xs text-neutral-500">
+                  Phase 2: OAuth Integration (Search Public + Premium Features)
+                </div>
+              </div>
+            </div>
+          </>
+        )}
 
         {/* Navigation */}
         <div className="mt-8 text-center">
