+++ components/search/QuickFilters/QuickFilters.tsx
@@ -110,29 +110,31 @@ export const QuickFilters: FC<QuickFiltersProps> = ({
           <ChevronDown className="absolute right-0 top-1/2 -translate-y-1/2 w-3 h-3 text-neutral-500 pointer-events-none" />
         </div>
 
-        {/* Category Filter */}
-        <div className="relative">
-          <select
-            id="category-filter"
-            name="category"
-            value={filters.category}
-            onChange={(e) => handleFilterChange('category', e.target.value)}
-            className="
-              appearance-none bg-transparent
-              text-sm font-medium text-neutral-700
-              pr-6 pl-1 py-1
-              border-none outline-none
-              cursor-pointer
-              hover:text-byggern-blue
-              transition-colors duration-150
-            "
-          >
-            {currentCategoryOptions.map(option => (
-              <option key={option} value={option}>{option}</option>
-            ))}
-          </select>
-          <ChevronDown className="absolute right-0 top-1/2 -translate-y-1/2 w-3 h-3 text-neutral-500 pointer-events-none" />
-        </div>
+        {/* Category Filter - Hidden for now, will be implemented properly later */}
+        {false && (
+          <div className="relative">
+            <select
+              id="category-filter"
+              name="category"
+              value={filters.category}
+              onChange={(e) => handleFilterChange('category', e.target.value)}
+              className="
+                appearance-none bg-transparent
+                text-sm font-medium text-neutral-700
+                pr-6 pl-1 py-1
+                border-none outline-none
+                cursor-pointer
+                hover:text-byggern-blue
+                transition-colors duration-150
+              "
+            >
+              {currentCategoryOptions.map(option => (
+                <option key={option} value={option}>{option}</option>
+              ))}
+            </select>
+            <ChevronDown className="absolute right-0 top-1/2 -translate-y-1/2 w-3 h-3 text-neutral-500 pointer-events-none" />
+          </div>
+        )}
 
 
       </div>
