--- docs/null-based-masking-implementation.md
@@ -1,95 +1,0 @@
-# ✅ Null-Based Data Masking Implementation Complete
-
-## 🎯 Successfully Updated to Backend Security-First Design
-
-### **Changes Made:**
-
-**1. Product Type Interface Updated**
-```typescript
-interface Product {
-  // Always present fields
-  id: string;
-  navn: string;
-  lagerstatus: 'På lager' | 'Utsolgt';  // Simplified to 2 states
-  // ... other always-visible fields
-  
-  // Security-filtered fields (null = unauthorized, value = authorized)
-  lagerantall: number | null;    // ← Changed from optional to null-based
-  grunnpris: number | null;      // ← Changed from optional to null-based  
-  nettopris: number | null;      // ← Changed from optional to null-based
-}
-```
-
-**2. Column Rendering Logic Updated**
-```typescript
-// Before: Authentication-based checks
-render: (value) => authState.isAuthenticated ? value : '****'
-
-// After: Null-based checks (backend security)
-render: (value) => value !== null ? value : '****'
-
-// Status column special case
-render: (value) => value === null ? <empty-space> : <StatusIndicator>
-```
-
-**3. API Transformation Updated**
-```typescript
-// Backend transformation ensures all required fields present
-lagerantall: backendProduct.lagerantall !== undefined ? backendProduct.lagerantall : null,
-grunnpris: backendProduct.grunnpris !== undefined ? backendProduct.grunnpris : null,
-nettopris: backendProduct.nettopris !== undefined ? backendProduct.nettopris : null,
-```
-
-**4. Mock Data Updated**
-Mixed data with both null (unauthenticated) and values (authenticated) for testing:
-- Product 1, 3: `lagerantall: null, grunnpris: null, nettopris: null`
-- Product 2, 4+: Actual values for authenticated state visualization
-
-### **✅ Benefits Achieved:**
-
-🔒 **Enterprise Security**: Backend filters sensitive data before transmission
-📏 **Consistent Layout**: Same table structure regardless of authorization  
-⚡ **Performance**: Single API response structure, frontend handles display
-🛡️ **Type Safety**: Predictable null checks instead of undefined
-🎯 **Simple Logic**: Clean `value !== null` checks for masking
-
-### **🔄 Backend Integration Ready:**
-
-**API Response Structure Expected:**
-```json
-{
-  "products": [
-    {
-      "id": "12345",
-      "navn": "Product Name",
-      "lagerstatus": "På lager",     // Always present
-      "lagerantall": null,           // null = show "****"  
-      "grunnpris": null,             // null = show "****"
-      "nettopris": null              // null = show "****"
-    }
-  ]
-}
-```
-
-**Display Logic:**
-- `null` values → Show `****` 
-- Actual values → Show formatted data
-- Status column `null` → Show empty space
-- Status column value → Show ● or × indicator
-
-### **✅ Validation Complete:**
-
-- ✅ **TypeScript**: All type errors resolved
-- ✅ **Runtime**: Application compiles and runs successfully  
-- ✅ **Logic**: Null-based masking logic implemented correctly
-- ✅ **Testing**: Mock data provides both authenticated/unauthenticated states
-- ✅ **Performance**: No authentication state dependencies in render logic
-
-### **🚀 Ready For:**
-
-1. **Backend Integration**: API can return null for unauthorized fields
-2. **Authentication Testing**: Frontend handles both null and value states
-3. **Production Deployment**: Security-first design implemented
-4. **User Experience**: Seamless masking/unmasking with proper authorization
-
-**The hybrid security-first + UX-optimized approach is now fully implemented and ready for backend integration!**
