+++ docs/plans/implement-catalog-stats-welcome-screen.md
